Move signed header errors to separate file
